+++ R/time_decompose.R
@@ -81,8 +81,6 @@
 #'
 #' library(dplyr)
 #'
-#' data(tidyverse_cran_downloads)
-#'
 #' # Basic Usage
 #' tidyverse_cran_downloads %>%
 #'     time_decompose(count, method = "stl")